+++ docs/CHANGELOG.md
@@ -15,6 +15,25 @@ All notable changes to this project will be documented in this file.
 ### Changed
 - .db extension changed to .pydbm
 
+### Fixed
+- Fix: Update obj on DB when updating the field on the instance. [#43](https://github.com/hakancelikdev/pydbm/pull/43)
+  ````python
+  import pydbm
+  
+  
+  class Model(pydbm.DbmModel):
+    username: str
+  
+  
+  model = Model(username="username")
+  model.save()
+  
+  model.username = "new_username"
+  model.save()
+  
+  assert Model.objects.get(pk=model.pk) == Model(username="new_username")
+  ````
+
 ## [0.4.0] - 2022-12-23
 
 ### Added